Moved app.module to app directory and added ng-bootstrap
