+++ app/main.ts
@@ -1,6 +1,6 @@
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic';
 
-import { AppModule } from '../app.module.js';
+import { AppModule } from './app.module.js';
 
 const platform = platformBrowserDynamic();
