+++ app/states.ts
@@ -1,0 +1,8 @@
+import {Home} from "./components/home";
+import {About} from "./components/about";
+import {Transition} from "ui-router-ng2";
+
+/** States */
+export const homeState = { name: 'home', url: '/home',  component: Home };
+
+export const aboutState = { name: 'about', url: '/about',  component: About };
