+++ index.html
@@ -6,7 +6,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1">
     <link rel="stylesheet" href="styles.css">
-
+    <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.css" />
     <!-- Polyfills for older browsers -->
     <script src="node_modules/core-js/client/shim.min.js"></script>
     <script src="node_modules/zone.js/dist/zone.js"></script>
@@ -14,7 +14,9 @@
     <script src="node_modules/systemjs/dist/system.src.js"></script>
     <script src="https://unpkg.com/typescript@1.8.10/lib/typescript.js"></script>
 
+
     <script src="systemjs.config.js"></script>
+
     <script>
       System.import('app').catch(function(err){ console.error(err); });
     </script>
