+++ package.json
@@ -28,9 +28,11 @@
     "@angular/platform-browser-dynamic": "~2.1.1",
     "@angular/router": "~3.1.1",
     "@angular/upgrade": "~2.1.1",
+    "@ng-bootstrap/ng-bootstrap": "^1.0.0-alpha.13",
     "angular-in-memory-web-api": "~0.1.13",
     "bluebird": "^3.4.6",
     "body-parser": "^1.15.2",
+    "bootstrap": "^3.3.7",
     "core-js": "^2.4.1",
     "express": "^4.14.0",
     "morgan": "^1.7.0",
@@ -48,6 +50,7 @@
     "concurrently": "^3.0.0",
     "grunt": "^0.4.5",
     "grunt-concurrent": "^2.3.1",
+    "grunt-contrib-clean": "^1.0.0",
     "grunt-contrib-copy": "^1.0.0",
     "grunt-contrib-watch": "^1.0.0",
     "grunt-ts": "^6.0.0-beta.3",
